ihm/server/requirements.txt
# FastAPI and server dependencies
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5

# Additional dependencies for functionality
python-multipart==0.0.21

# Database and cache dependencies for local development
psycopg2-binary==2.9.11
redis==7.1.0

# Environment and configuration
python-dotenv==1.2.1

# MQTT client for AI Assistant agent communication
paho-mqtt==2.1.0

# REST API client
httpx==0.28.1
